+++ App.js
@@ -2,13 +2,15 @@ import { AppLoading } from 'expo';
 import { Asset } from 'expo-asset';
 import * as Font from 'expo-font';
 import React, { useState } from 'react';
-import { Platform, StatusBar, StyleSheet, Text, View } from 'react-native';
+import { Platform, StatusBar, StyleSheet, Text, View, SafeAreaView } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 
 import AppNavigator from './navigation/AppNavigator'
 
 export default function App(props) {
   const [isLoadingComplete, setLoadingComplete] = useState(false);
+  const platformVersion = parseInt(Platform.Version, 10);
+
   if (!isLoadingComplete && !props.skipLoadingScreen) {
     return (
       <AppLoading
@@ -19,10 +21,13 @@ export default function App(props) {
     );
   } else {
     return (
-      <View style={styles.container}>
-        {Platform.OS === 'ios' && <StatusBar barStyle="default" />}
+      <SafeAreaView 
+        style={styles.container}
+      >
+        {Platform.OS === 'ios' && <StatusBar backgroundColor="blue" barStyle="light-content" />}
+        {Platform.OS === 'android' && <StatusBar backgroundColor="blue" barStyle="light-content" />}
         <AppNavigator />
-      </View>
+      </SafeAreaView>
     );
   }
 }